sphinx-agent: Move directives types to a sub-module
