+++ lib/esbonio/esbonio/server/features/directives/__init__.py
@@ -3,24 +3,14 @@ from __future__ import annotations
 import inspect
 import typing
 
-import attrs
+from lsprotocol import types as lsp
 
 from esbonio import server
+from esbonio.sphinx_agent import types
 
 if typing.TYPE_CHECKING:
     from collections.abc import Coroutine
     from typing import Any
-
-
-@attrs.define
-class Directive:
-    """Represents a directive."""
-
-    name: str
-    """The name of the directive, as the user would type in an rst file."""
-
-    implementation: str | None
-    """The dotted name of the directive's implementation."""
 
 
 class DirectiveProvider:
@@ -28,7 +18,9 @@ class DirectiveProvider:
 
     def suggest_directives(
         self, context: server.CompletionContext
-    ) -> list[Directive] | None | Coroutine[Any, Any, list[Directive] | None]:
+    ) -> (
+        list[types.Directive] | None | Coroutine[Any, Any, list[types.Directive] | None]
+    ):
         """Given a completion context, suggest directives that may be used."""
         return None
 
@@ -57,7 +49,7 @@ class DirectiveFeature(server.LanguageFeature):
 
     async def suggest_directives(
         self, context: server.CompletionContext
-    ) -> list[Directive]:
+    ) -> list[types.Directive]:
         """Suggest directives that may be used, given a completion context.
 
         Parameters
@@ -65,11 +57,11 @@ class DirectiveFeature(server.LanguageFeature):
         context
            The completion context.
         """
-        items: list[Directive] = []
+        items: list[types.Directive] = []
 
         for provider in self._providers.values():
             try:
-                result: list[Directive] | None = None
+                result: list[types.Directive] | None = None
 
                 aresult = provider.suggest_directives(context)
                 if inspect.isawaitable(aresult):
